It compiles and runs on the device!
One weird thing that I have yet to explain: when the program terminates,
we get '?SYNTAX ERROR IN 48384' on the screen, and instead of being
dropped into the system monitor prompt, we get to a BASIC prompt.

Maybe it's to do with the "missing" parts of our executable, since we're
emitting just the cpu instructions and nothing before/after them. When
we were using cc65 to compile C programs, there was a bunch of junk in
the linker script -- maybe that had something to do with exiting
"cleanly"?